1) Modified pilotAnalysis2.m so that the user does not have to click "Edit" upon entering a new trial. This saves some mouse clicks.
2) Modified pilotAnalysis2.m and SEQ_GUI.m so that audioplayer is used in
lieu of soundsc/wavplay on a Mac computer. There is still some issues in
it. For exapmle, the "play" button can be used only once. Beyond the first
time, sound won't be played. At least this was the case on my Windows
machine running audioplayer. I don't know whether this will be the same on
a Mac, because I don't have my hands on a Mac computer. Maybe who ever
watns to use Mac can revise the coding and make it work better.