Modifications after the first IRR meeting
1. Added new option: "redoCatOnly": allows the redoing of the categorical
labels (accuracy and fluency) for specified trial(s). The usage is similar
to the "redo" option,e.g.,

pilotAnalysis2('fmrih', 3, 'SEQ03C07_behavioral', 'test8', 3, 'scai',
'noFilter', 'redoCatOnly', [1, 18, 22, 27, 2])

The above command allows the user to redo the categorical ratings for
the five specified trials.

Note that 'recoCatOnly' cannot be used in conjuction with 'redo'.

2. Let the categorical rating window be accompanied by a spectrogram
window to help the user judge things like the presence or absence of the
initial fricative in words such as "FTHAMCH"

3. Note that a new script is added (show_spectrogram.m) and it needs to be
downloaded for the new version to be functional.