Added a new option to the GUI that allows the labeling the onset of the "starter" behavior before the onset of the word.
The first GUI window now has a check box "Starter", which if checked, will
cause the second GUI to let the user input three time labels, instead of
the default two. The three time labels are the starter onset, word onset
and word end, in that order. The program checks the correctness of the
ordering of these three labels.

A new field is added to data{ii} to indicate whether starter behavioral
exists in the trial or not (as inputted by the user). The name of this new
field is "bStarter". If bStarter==1, data{ii} will contain a field called
"starterOnset" for the onset time of the starter behavior (as inputted by
the user). Otherwise the starterOnset field will not exist in the
particular data cell.